Moved rest of the methods and properties to new viewModel
