+++ src/ReferenceAnalyzer.UI/Views/SolutionView.xaml.cs
@@ -22,12 +22,24 @@ namespace ReferenceAnalyzer.UI.Views
                        viewModel => viewModel.Path,
                        view => view.Path.Text)
                    .DisposeWith(disposableRegistration);
+
+
+                this.OneWayBind(ViewModel,
+                        viewModel => viewModel.LastSolutions,
+                        view => view.LastLoadedSolutions.Items)
+                    .DisposeWith(disposableRegistration);
+
+                this.BindCommand(ViewModel,
+                    viewModel => viewModel.PickSolutionFile,
+                    view => view.PickSolutionLocation)
+                .DisposeWith(disposableRegistration);
             });
 
             AvaloniaXamlLoader.Load(this);
         }
 
         public TextBox Path => this.FindControl<TextBox>(nameof(Path));
-
+        public Button PickSolutionLocation => this.FindControl<Button>(nameof(PickSolutionLocation));
+        public ListBox LastLoadedSolutions => this.FindControl<ListBox>(nameof(LastLoadedSolutions));
     }
 }
